added bit masks

+++ src/Commands/MoveCommand.cpp
@@ -15,7 +15,7 @@ uint16_t MoveCommand::NumberOfArguments()
     return 2;
 }
 
-uint64_t MoveCommand::Execute(uint16_t instruction, std::vector<uint8_t> additionalWords, std::vector<uint8_t> &Registers, std::vector<uint8_t> &SpecialRegisters, uint16_t ProgramCounter, uint16_t& StackPointer)
+uint64_t MoveCommand::Execute(uint16_t instruction, std::vector<uint16_t> additionalWords, std::vector<uint8_t> &Registers, std::vector<uint8_t> &SpecialRegisters, uint16_t ProgramCounter, uint16_t& StackPointer)
 {
 
     return ProgramCounter +1;
